Look for one of two different session cookie names
The name of the session cookie changed at some point. This PR updates
the cookie-fetching logic to look for one of two different cookie
named. I'm not sure if the cookie names have changed across the board or
if the names have only changed for some users as part of the SSOe
rollout.